Move yAxis so it can be shared
