+++ src/components/LineChart/components/index.ts
@@ -3,4 +3,3 @@ export {Legend} from './Legend';
 export {Line} from './Line';
 export {Tooltip} from './Tooltip';
 export {XAxis} from './XAxis';
-export {YAxis} from './YAxis';
