+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export {LineChart} from './LineChart';
 export {NormalizedStackedBar} from './NormalizedStackedBar';
 export {Sparkline} from './Sparkline';
+export {YAxis} from './YAxis';
